arch2_microservices/borrowing_service/requirements.txt
grpcio
grpcio-tools
pytest

sqlalchemy 
psycopg2-binary